requirements.txt
python = 3.7.7
pygame = 1.9.6
pytorch = 1.6.0
torchvision = 0.7.0
numpy = 1.19.1